- Corrected bug where parent fields where never ignored in AbstractSyncManager. - Corrected bug when comparing SyncModels for equality. - Corrected encryption implementation to deal only with bytes instead of strings.

+++ src/main/java/br/com/estudio89/syncing/models/SyncModel.java
@@ -64,7 +64,7 @@ public abstract class SyncModel<T extends SyncModel<?>> extends SugarRecord<T> {
             return false;
         }
 
-        return ((SyncModel<?>) o).getId() == this.getId();
+        return ((SyncModel<?>) o).getId().longValue() == this.getId().longValue();
     }
 
     private static <Model extends  SyncModel> Field getDateField(Class<Model> type) {
